requirements.txt
python-telegram-bot==12.7
groq
psutil